+++ dd-sdk-android-gradle-plugin/src/main/kotlin/com/datadog/gradle/plugin/DdAndroidGradlePlugin.kt
@@ -48,6 +48,10 @@ class DdAndroidGradlePlugin @Inject constructor(
 
     /** @inheritdoc */
     override fun apply(target: Project) {
+        check(CurrentAgpVersion.CAN_USE_PLUGIN) {
+            "Minimum supported AGP version is 8.0.0, Datadog Gradle Plugin cannot be applied."
+        }
+
         val extension = target.extensions.create(EXT_NAME, DdExtension::class.java)
         val apiKeyProvider = resolveApiKey(target)
         // need to use withPlugin instead of afterEvaluate, because otherwise generated assets